refactor: extract filter into its own component for code cleanliness and use Industry enum
